Split format-specific tests into test-formats
Summary:
Move CSV and Parquet tests into a dedicated `cli_formats_tests.py` file,
run via `make test-formats`. These are format-specific tests that exercise
parsers (CSV separators, Parquet chunking, etc.) and don't belong in
generic CLI or training test suites.

- `test-cli`: CLI interface tests (flags, arg parsing, compress/decompress)
- `test-formats`: format-specific tests (CSV, Parquet)
- `test-train`: generic training infrastructure (SDDL2, u8, ML, inline)

The CI job runs both `test-train` and `test-formats` together.

Test Plan:
```
make test-cli
make test-formats
make test-train
```
All pass independently.
