+++ cli/tests/cli_train_tests.py
@@ -17,7 +17,6 @@ import unittest
 
 import command_utils
 from abstract_compression_test import (
-    _CsvBaseTest,
     _MLBaseTest,
     _TrainBaseTest,
     _TrainInlineBaseTest,
@@ -31,120 +30,6 @@ from command_utils import (
     execute_train,
 )
 from file_utils import file_contents_match, input_dir_path, profile_dir_path
-
-
-class CsvGreedyTest(_CsvBaseTest):
-    """
-    Test case for CSV training and compression using the greedy trainer.
-
-    This test demonstrates the train-compress-decompress workflow for CSV files
-    using the greedy training algorithm. The greedy trainer optimizes compression
-    by making locally optimal choices at each step.
-
-    Sample files are located in cli/tests/sample_files/csv/
-    Output files are stored in {output_dir_path}
-    """
-
-    @property
-    def trainer_name(self) -> str | None:
-        return "greedy"
-
-    def test_train_compress_decompress(self):
-        """
-        Test the train, compress, and decompress workflow for CSV files using the greedy trainer.
-
-        This test:
-        1. Trains a compressor on the CSV files in cli/tests/sample_files/csv/ using the greedy trainer
-        2. Saves the trained compressor to {output_dir_path}/trained_compressor.zlc
-        3. Uses the trained compressor to compress and decompress the CSV files
-        4. Verifies that the decompressed files match the originals
-        """
-        self.train_compress_decompress()
-
-
-class CsvSaveAceStateTest(_CsvBaseTest):
-    """
-    Test case for CSV training with --save-ace-state flag.
-
-    Verifies that training with --save-ace-state produces a compressor
-    that correctly compresses and decompresses files.
-    """
-
-    def test_train_compress_decompress(self):
-        execute_train(
-            compressor_info=self.training_compressor_info,
-            uncompressed_dir=input_dir_path(self.input_dir_name),
-            trained_compressor_path=self.compressor_info.compressor_str,
-            extra_args="--save-ace-state",
-        )
-        self.compress_and_decompress_samples()
-
-
-class CsvFullSplitTest(_CsvBaseTest):
-    """
-    Test case for CSV training and compression using the full-split trainer.
-
-    This test demonstrates the train-compress-decompress workflow for CSV files
-    using the full-split training algorithm. The full-split trainer optimizes compression
-    by analyzing the entire dataset before making decisions.
-
-    Sample files are located in cli/tests/sample_files/csv/
-    Output files are stored in {output_dir_path}/csv_full_split/
-    """
-
-    @property
-    def trainer_name(self) -> str | None:
-        return "full-split"
-
-    def test_train_compress_decompress(self):
-        """
-        Test the train, compress, and decompress workflow for CSV files using the full-split trainer.
-
-        This test:
-        1. Trains a compressor on the CSV files in cli/tests/sample_files/csv/ using the full-split trainer
-        2. Saves the trained compressor to {output_dir_path}/trained_compressor.zlc
-        3. Uses the trained compressor to compress and decompress the CSV files
-        4. Verifies that the decompressed files match the originals
-        """
-        self.train_compress_decompress()
-
-
-class CsvBottomUpTest(_CsvBaseTest):
-    """
-    Test case for CSV training and compression using the bottom-up trainer.
-
-    Sample files are located in cli/tests/sample_files/csv/
-    Output files are stored in {output_dir_path}
-    """
-
-    @property
-    def trainer_name(self) -> str | None:
-        return "bottom-up"
-
-    def test_train_compress_decompress(self):
-        """
-        Test the train, compress, and decompress workflow for CSV files using the bottom-up trainer.
-        """
-        self.train_compress_decompress()
-
-
-class CsvChunkedTest(_CsvBaseTest):
-    """
-    Test case for CSV training and compression using the trainer with chunking.
-
-    Sample files are located in cli/tests/sample_files/csv/
-    Output files are stored in {output_dir_path}
-    """
-
-    @property
-    def extra_args(self) -> str | None:
-        return "--chunk-size 1M"
-
-    def test_train_compress_decompress(self):
-        """
-        Test the train, compress, and decompress workflow for CSV files with chunking.
-        """
-        self.train_compress_decompress()
 
 
 class Sddl2ChunkedTrainTest(unittest.TestCase):
@@ -374,95 +259,6 @@ class MLSelectorTest(_MLBaseTest):
         4. Verifies that the decompressed files match the originals
         """
         self.train_compress_decompress()
-
-
-class ParquetTest(_TrainBaseTest):
-    """
-    Parquet compression tests with training.
-
-    """
-
-    @property
-    def input_dir_name(self) -> str:
-        return "parquet"
-
-    @property
-    def compressor_profile_name(self) -> str:
-        return "parquet"
-
-    def test_train_compress_decompress(self):
-        """
-        Test the train, compress, and decompress workflow for Parquet files using the clustering trainer.
-        """
-        self.train_compress_decompress()
-
-
-class ParquetChunkedTest(ParquetTest):
-    """
-    Verify that --chunk-size is wired through to the parquet profile.
-
-    Trains twice — once with the default chunk size and once with an explicit
-    --chunk-size 2M — then checks that the serialized compressors differ
-    (the chunk size is baked into the profile) and that both round-trip
-    correctly.
-    """
-
-    def test_chunk_size_affects_output(self):
-        profile = CompressorInfo(
-            compressor_str=self.compressor_profile_name,
-            compressor_type=CompressorType.PROFILE,
-        )
-
-        default_trained = os.path.join(self.output_dir_path, "default.zlc")
-        execute_train(
-            compressor_info=profile,
-            uncompressed_dir=input_dir_path(self.input_dir_name),
-            trained_compressor_path=default_trained,
-        )
-
-        chunked_trained = os.path.join(self.output_dir_path, "chunked.zlc")
-        execute_train(
-            compressor_info=profile,
-            uncompressed_dir=input_dir_path(self.input_dir_name),
-            trained_compressor_path=chunked_trained,
-            extra_args="--chunk-size 2M",
-        )
-
-        with open(default_trained, "rb") as f:
-            default_bytes = f.read()
-        with open(chunked_trained, "rb") as f:
-            chunked_bytes = f.read()
-        self.assertNotEqual(
-            default_bytes,
-            chunked_bytes,
-            "Trained compressors should differ when --chunk-size is changed",
-        )
-
-        sample = self.input_samples[0]
-        for trained_path, label in [
-            (default_trained, "default"),
-            (chunked_trained, "chunked"),
-        ]:
-            trained = CompressorInfo(
-                compressor_str=trained_path,
-                compressor_type=CompressorType.FILE,
-            )
-            compressed = sample.compressed_file_path + f".{label}"
-            execute_compress(
-                file_to_compress_path=sample.orig_file_path,
-                compressor_info=trained,
-                compressed_file_path=compressed,
-                extra_args=None,
-            )
-            decompressed = compressed + ".out"
-            execute_decompress(
-                compressed_file_path=compressed,
-                decompressed_file_path=decompressed,
-            )
-            self.assertTrue(
-                file_contents_match(sample.orig_file_path, decompressed),
-                f"Round-trip failed for {label} path",
-            )
 
 
 class U16TrainInlineTest(_TrainInlineBaseTest):
